requirements.txt
fastapi==0.128.0
uvicorn==0.25.0
motor==3.7.1
passlib==1.7.4
pydantic==2.12.5
PyJWT==2.10.1
python-dotenv==1.0.0
reportlab==4.4.9
openpyxl==3.1.2
email-validator
requests